address bulk publish review comments in #5256
